requirements.txt
fastapi
uvicorn
python-multipart
librosa
numpy
pillow
soundfile